Reduce OAS generation boilerplate
* Create a common list schema from a given module with schema
* Create macro to reduce boilerplate for object schemas
