+++ lib/logflare_web/open_api_schemas.ex
@@ -2,175 +2,97 @@ defmodule LogflareWeb.OpenApiSchemas do
   alias OpenApiSpex.Schema
 
   defmodule Endpoint do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{
-      type: :object,
-      properties: %{
-        token: %Schema{type: :string},
-        name: %Schema{type: :string},
-        query: %Schema{type: :string},
-        source_mapping: %Schema{type: :object},
-        sandboxable: %Schema{type: :boolean},
-        cache_duration_seconds: %Schema{type: :integer},
-        proactive_requerying_seconds: %Schema{type: :integer},
-        max_limit: %Schema{type: :integer},
-        enable_auth: %Schema{type: :boolean},
-        inserted_at: %Schema{type: :string, format: :"date-time"},
-        updated_at: %Schema{type: :string, format: :"date-time"}
-      },
-      required: [:name, :query]
-    })
-
-    def response(), do: {"Endpoint Response", "application/json", __MODULE__}
-    def params(), do: {"Endpoint Params", "application/json", __MODULE__}
+    @properties %{
+      token: %Schema{type: :string},
+      name: %Schema{type: :string},
+      query: %Schema{type: :string},
+      source_mapping: %Schema{type: :object},
+      sandboxable: %Schema{type: :boolean},
+      cache_duration_seconds: %Schema{type: :integer},
+      proactive_requerying_seconds: %Schema{type: :integer},
+      max_limit: %Schema{type: :integer},
+      enable_auth: %Schema{type: :boolean},
+      inserted_at: %Schema{type: :string, format: :"date-time"},
+      updated_at: %Schema{type: :string, format: :"date-time"}
+    }
+    use LogflareWeb.OpenApi, properties: @properties, required: [:name, :query]
   end
 
   defmodule Notification do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{
-      type: :object,
-      properties: %{
-        team_user_ids_for_email: %Schema{type: :array, allOf: %Schema{type: :string}},
-        team_user_ids_for_sms: %Schema{type: :array, allOf: %Schema{type: :string}},
-        team_user_ids_for_schema_updates: %Schema{type: :array, allOf: %Schema{type: :string}},
-        other_email_notifications: %Schema{type: :string},
-        user_email_notifications: %Schema{type: :boolean},
-        user_text_notifications: %Schema{type: :boolean},
-        user_schema_update_notifications: %Schema{type: :boolean}
-      }
-    })
-
-    def response(), do: {"Notification Response", "application/json", __MODULE__}
+    @properties %{
+      team_user_ids_for_email: %Schema{type: :array, allOf: %Schema{type: :string}},
+      team_user_ids_for_sms: %Schema{type: :array, allOf: %Schema{type: :string}},
+      team_user_ids_for_schema_updates: %Schema{type: :array, allOf: %Schema{type: :string}},
+      other_email_notifications: %Schema{type: :string},
+      user_email_notifications: %Schema{type: :boolean},
+      user_text_notifications: %Schema{type: :boolean},
+      user_schema_update_notifications: %Schema{type: :boolean}
+    }
+    use LogflareWeb.OpenApi, properties: @properties, required: []
   end
 
   defmodule Source do
-    require OpenApiSpex
+    @properties %{
+      name: %Schema{type: :string},
+      token: %Schema{type: :string},
+      id: %Schema{},
+      favorite: %Schema{type: :boolean},
+      webhook_notification_url: %Schema{type: :string},
+      api_quota: %Schema{type: :integer},
+      slack_hook_url: %Schema{type: :string},
+      bigquery_table_ttl: %Schema{type: :integer},
+      public_token: %Schema{type: :string},
+      bq_table_id: %Schema{type: :string},
+      bq_table_schema: %Schema{type: :object},
+      has_rejected_events: %Schema{type: :boolean},
+      metrics: %Schema{type: :object},
+      notifications: %Schema{type: :array, items: Notification},
+      custom_event_message_keys: %Schema{type: :string},
+      inserted_at: %Schema{type: :string, format: :"date-time"},
+      updated_at: %Schema{type: :string, format: :"date-time"}
+    }
 
-    OpenApiSpex.schema(%{
-      type: :object,
-      properties: %{
-        name: %Schema{type: :string},
-        token: %Schema{type: :string},
-        id: %Schema{},
-        favorite: %Schema{type: :boolean},
-        webhook_notification_url: %Schema{type: :string},
-        api_quota: %Schema{type: :integer},
-        slack_hook_url: %Schema{type: :string},
-        bigquery_table_ttl: %Schema{type: :integer},
-        public_token: %Schema{type: :string},
-        bq_table_id: %Schema{type: :string},
-        bq_table_schema: %Schema{type: :object},
-        has_rejected_events: %Schema{type: :boolean},
-        metrics: %Schema{type: :object},
-        notifications: %Schema{type: :array, items: Notification},
-        custom_event_message_keys: %Schema{type: :string},
-        inserted_at: %Schema{type: :string, format: :"date-time"},
-        updated_at: %Schema{type: :string, format: :"date-time"}
-      },
-      required: [:name]
-    })
-
-    def response(), do: {"Source Response", "application/json", __MODULE__}
-    def params(), do: {"Source Params", "application/json", __MODULE__}
+    use LogflareWeb.OpenApi, properties: @properties, required: [:name]
   end
 
   defmodule User do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{
-      type: :object,
-      properties: %{
-        email: %Schema{type: :string},
-        provider: %Schema{type: :string},
-        api_key: %Schema{type: :string},
-        email_preferred: %Schema{type: :string},
-        name: %Schema{type: :string},
-        image: %Schema{type: :string},
-        email_me_product: %Schema{type: :boolean},
-        phone: %Schema{type: :string},
-        bigquery_project_id: %Schema{type: :string},
-        bigquery_dataset_location: %Schema{type: :string},
-        bigquery_dataset_id: %Schema{type: :string},
-        api_quota: %Schema{type: :integer},
-        company: %Schema{type: :string},
-        token: %Schema{type: :string}
-      },
-      required: [:name]
-    })
-
-    def response(), do: {"User Response", "application/json", __MODULE__}
+    @properties %{
+      email: %Schema{type: :string},
+      provider: %Schema{type: :string},
+      api_key: %Schema{type: :string},
+      email_preferred: %Schema{type: :string},
+      name: %Schema{type: :string},
+      image: %Schema{type: :string},
+      email_me_product: %Schema{type: :boolean},
+      phone: %Schema{type: :string},
+      bigquery_project_id: %Schema{type: :string},
+      bigquery_dataset_location: %Schema{type: :string},
+      bigquery_dataset_id: %Schema{type: :string},
+      api_quota: %Schema{type: :integer},
+      company: %Schema{type: :string},
+      token: %Schema{type: :string}
+    }
+    use LogflareWeb.OpenApi,
+      properties: @properties,
+      required: [:email, :provider, :token, :provider_uid, :api_key]
   end
 
   defmodule TeamUser do
-    require OpenApiSpex
+    @properties %{
+      email: %Schema{type: :string},
+      name: %Schema{type: :string}
+    }
 
-    OpenApiSpex.schema(%{
-      type: :object,
-      properties: %{
-        email: %Schema{type: :string},
-        name: %Schema{type: :string}
-      }
-    })
-
-    def response(), do: {"Team Response", "application/json", __MODULE__}
+    use LogflareWeb.OpenApi, properties: @properties, required: [:email, :name]
   end
 
   defmodule Team do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{
-      type: :object,
-      properties: %{
-        name: %Schema{type: :string},
-        token: %Schema{type: :string},
-        user: User,
-        team_users: %Schema{type: :array, items: TeamUser}
-      },
-      required: [:name]
-    })
-
-    def response(), do: {"Team Response", "application/json", __MODULE__}
-    def params(), do: {"Team Params", "application/json", __MODULE__}
-  end
-
-  defmodule EndpointList do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{type: :array, items: Endpoint})
-    def response(), do: {"Endpoint List Response", "application/json", __MODULE__}
-  end
-
-  defmodule SourceList do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{type: :array, items: Source})
-    def response(), do: {"Source List Response", "application/json", __MODULE__}
-  end
-
-  defmodule TeamList do
-    require OpenApiSpex
-
-    OpenApiSpex.schema(%{type: :array, items: Team})
-    def response(), do: {"Team List Response", "application/json", __MODULE__}
-  end
-
-  defmodule Created do
-    def response(schema), do: {"Created Response", "application/json", schema}
-  end
-
-  defmodule Accepted do
-    require OpenApiSpex
-    OpenApiSpex.schema(%{})
-
-    def response(), do: {"Accepted Response", "text/plain", __MODULE__}
-  end
-
-  defmodule NotFound do
-    require OpenApiSpex
-    OpenApiSpex.schema(%{})
-
-    def response(), do: {"Not found", "text/plain", __MODULE__}
+    @properties %{
+      name: %Schema{type: :string},
+      token: %Schema{type: :string},
+      user: User,
+      team_users: %Schema{type: :array, items: TeamUser}
+    }
+    use LogflareWeb.OpenApi, properties: @properties, required: [:name]
   end
 end
